refactor(api): thin the service graph — execution engine owns its seams
Dependency-graph audit (SOLID/SoC) and the fixes it demanded:

- the run-execution engine (RunExecutionService, RunsWorkerService,
  RunStreamBridgeService) moves from chats/ into runs/ — it was pulling
  queue, models, compaction, and titles into the chats module while
  being runs-domain logic
- new RunWorkerModule: the bootable execution module (queue consumers +
  executor + bridge + dispatch), imported by the api today and by the
  dedicated worker entrypoint when #116 lands
- new RunDispatchService: the publish seam — queue names, job payload
  shapes, per-run deadman scheduling, ensure-queues-once, and the
  fail-the-run-on-enqueue-failure contract all live behind one
  dispatch() call. chats no longer imports queue/ AT ALL; chat-loop
  sheds its queueReady state and enqueue mechanics
- @Global DbModule provides TenantDbService once (five modules each
  re-provided their own instance)
- chat-domain helpers return home: isCompletedAssistantTurn →
  chats-repository, CHAT_SYSTEM_PROMPT → context-builder

Resulting cross-feature edges: queue is consumed only by runs; chats →
auth/db/models/runs (was seven targets); compaction/titles → chats only
for repositories and pure context fns; no module cycles. Boundary rules
recorded in apps/api/AGENTS.md.

Verified: tsgo + oxlint + prettier clean, unit 125, nest build +
openapi (module graph boots), full rls-test.sh green (RLS 11, queue 7,
e2e 33).
