+++ apps/api/src/chats/chats.module.ts
@@ -1,39 +1,24 @@
 import { Module } from '@nestjs/common';
 import { AuthModule } from '../auth/auth.module';
-import { TenantDbService } from '../db/tenant-db.service';
 import { ModelsModule } from '../models/models.module';
+import { RunWorkerModule } from '../runs/run-worker.module';
+import { RunsModule } from '../runs/runs.module';
 import { ChatLoopService } from './chat-loop.service';
 import { ChatsController } from './chats.controller';
 import { ChatsService } from './chats.service';
-import { QueueModule } from '../queue/queue.module';
-import { RunsModule } from '../runs/runs.module';
-import { CompactionModule } from '../compaction/compaction.module';
-import { TitlesModule } from '../titles/titles.module';
-import { RunExecutionService } from './run-execution.service';
-import { RunStreamBridgeService } from './run-stream-bridge';
-import { RunsWorkerService } from './runs-worker.service';
 
 // HTTP endpoints are safe to expose only because SessionAuthGuard derives the tenant
 // identity from a verified session. Controllers must never accept ownerUserId from
 // client input; that would recreate the #61 tenant-impersonation IDOR.
+//
+// Boundary: chats owns the turn (validate, persist message + run, supersede);
+// everything run-execution-shaped comes from RunWorkerModule (dispatch seam +
+// stream bridge) and RunsModule (abort registry) — chats knows nothing about
+// queues, workers, compaction, or titling.
 @Module({
-  imports: [
-    AuthModule,
-    ModelsModule,
-    QueueModule,
-    RunsModule,
-    CompactionModule,
-    TitlesModule,
-  ],
+  imports: [AuthModule, ModelsModule, RunsModule, RunWorkerModule],
   controllers: [ChatsController],
-  providers: [
-    TenantDbService,
-    ChatsService,
-    ChatLoopService,
-    RunExecutionService,
-    RunStreamBridgeService,
-    RunsWorkerService,
-  ],
+  providers: [ChatsService, ChatLoopService],
   exports: [ChatsService],
 })
 export class ChatsModule {}
